requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2024.1
Django==5.0.2
django-crispy-forms==2.1
django-recaptcha==4.0.0
django-tinymce==3.7.1
fontawesomefree==6.5.1
pillow==10.2.0
shortuuid==1.0.11
sqlparse==0.4.4
tzdata==2024.1
